moving files around; linting
