+++ .gitignore
@@ -11,3 +11,4 @@ __pycache__/
 *.pyc
 
 *.ipynb
+testing/
