+++ pyproject.toml
@@ -1,0 +1,15 @@
+[tool.poetry]
+name = "minlizhash"
+version = "0.1.0"
+description = "A simple MinHash LSH Implimentation from scratch"
+authors = ["Liz Goodwin <work@lizgood.win>"]
+license = "MIT"
+readme = "README.md"
+
+[tool.poetry.dependencies]
+python = "^3.12"
+
+
+[build-system]
+requires = ["poetry-core"]
+build-backend = "poetry.core.masonry.api"
