--- testplayground.py
@@ -1,14 +1,0 @@
-test_strings = [
-    "The quick brown fox jumps over the lazy dog. The dog was too lazy to react. The fox had a great time. In the distance, a bird chirped. The sun was setting, casting long shadows across the landscape. The fox looked around, its eyes gleaming in the fading light.",
-    "The lazy dog decided to sleep all day. The quick brown fox was too busy jumping around to notice. The grass was green and lush, a perfect bed for a lazy dog. The dog yawned and stretched, then settled down for a nap. The fox continued to jump around, oblivious to the dog's slumber.",
-    "The fox and the dog are good friends. They spend their days playing and jumping around. The forest was their playground, full of interesting smells and sights. They chased each other through the trees, their laughter echoing through the forest. At the end of the day, they would lie down side by side, content and tired.",
-    "The fox and the dog are good friends. hey spend their days playing heir playground, full of interesting smells and sights. They chased each other throughtired.",
-    "The fox and the are good friends. hey spend their days playing heir playground, full of interesting smells and sights. They chased each other throughtired.",
-    "In the heart of the ancient city, there stood a tall and slender tower. From its highest point, one could look out across the sprawling landscape, where buildings of all shapes and sizes jostled for space amidst the lush greenery of the surrounding parks and gardens.",
-    "In the heao whhhhat the f the ancient city, there stood a tall and slender tower. From its highest point, one could look out across the sprawling landscape, where buildings of all shapes and sizes jostled for space amidst the lush greenery of the surrounding parks and gardens.",
-    "The sun was setting, casting long shadows across the bustling city. The streets were filled with people, all hurrying to their various destinations. The air was filled with the sounds of chatter, laughter, and the occasional honk of a car horn.",
-    "In the quiet of the forest, the only sounds were the gentle rustling of leaves in the wind and the distant call of a bird. The air was fresh and clean, filled with the scent of damp earth and the faint hint of pine.",
-    "The ocean stretched out as far as the eye could see, its surface sparkling in the sunlight. Waves crashed against the shore, leaving behind a trail of white foam on the golden sand. Seagulls circled overhead, their cries echoing in the salty air.",
-    "The water bitch ass stretched out as far as the eye could see, its surface sparkling in the sunlight. Waves crashed against the shore, leaving behind a trail of white foam on the golden sand. Seagulls circled overhead, their cries echoing in the salty air.",
-    "The city was a hive of activity. Cars zoomed down the streets, their headlights cutting through the darkness. People walked briskly, their breath visible in the cold night air. The bright neon signs of shops and restaurants illuminated the sidewalks, casting colorful reflections on the wet pavement.",
-]
